Move AddrInfo from wrr to ringhash
